feat: implement logger
